Make various changes to support new function is_gisdk_named_array()
process_gisdk_result() now checks to see if the data object is of the particular
format returned when a GISDK named array comes over COM. If it is, then a
call to convert_to_named_list() will convert it. Several functions had to be
adjusted due to this change. The vignette and tests were also updated.